ingestion/requirements.txt
kafka-python==2.0.2
sseclient-py==1.8.0
requests==2.31.0
pyyaml>=6.0
